+++ packages/kit/test/apps/basics/test/test.js
@@ -1504,3 +1504,31 @@ test.describe.serial('Cookies API', () => {
 		expect(await span.innerText()).toContain('undefined');
 	});
 });
+
+test.describe('Serialization', () => {
+	test('A custom data type can be serialized/deserialized', async ({ page, clicknav }) => {
+		await page.goto('/serialization-basic');
+		expect(await page.textContent('h1')).toBe('It works!');
+
+		await clicknav('[href="/serialization-basic/child"]');
+		expect(await page.textContent('h1')).toBe('Client-side navigation also works!');
+	});
+
+	test('A custom data type can be serialized/deserialized on POST', async ({ page }) => {
+		await page.goto('/serialization-form');
+		await page.click('button');
+		expect(await page.textContent('h1')).toBe('It works!');
+
+		// Test navigating to the basic page works as intended
+		await page.locator('a').first();
+		expect(await page.textContent('h1')).toBe('It works!');
+	});
+
+	test('A custom data type can be serialized/deserialized on POST with use:enhance', async ({
+		page
+	}) => {
+		await page.goto('/serialization-form2');
+		await page.click('button');
+		expect(await page.textContent('h1')).toBe('It works!');
+	});
+});